* PlaylistLoader rewrite, not complete. Some regressions, ie crashes if you delete stuff during a load operation. Should work better when done * Introduced amarok.h, which defines some namespacing stuff and some useful functions. app.h is now hardly included at all. * Actually show engine status messages * Fix timeDisplay not updating when you click it in PlayerWindow * Big speed up to playlist scrolling. Don't double buffer everything, just flashy stuff. I noticed amaroK couldn't keep up with my scrolling. Now it can. * Some missed merges for EngineController
svn path=/trunk/kdeextragear-1/amarok/; revision=334029

+++ src/sliderwidget.cpp
@@ -113,7 +113,7 @@ void
 amaroK::PrettySlider::paintEvent( QPaintEvent* )
 {
     const int w   = orientation() == Qt::Horizontal ? width() : height();
-    const int pos = double((w-2) * Slider::value()) / maxValue();
+    const int pos = int(double((w-2) * Slider::value()) / maxValue());
     const int h   = THICKNESS;
 
     QPixmap  buf( size() );
